refactor: 리펙토링 및 디렉토리 이동
