+++ src/main/java/handler/LogoutHandler.java
@@ -3,7 +3,8 @@ package handler;
 import http.HttpRequest;
 import http.HttpResponse;
 import http.ResponseValueSetter;
-import repository.SessionManager;
+import session.SessionManager;
+import session.Session;
 import util.RequestContext;
 
 public class LogoutHandler extends MyHandler {
