+++ src/main/java/handler/StaticHandler.java
@@ -8,6 +8,7 @@ import http.startline.RequestLine;
 import java.io.IOException;
 import model.User;
 import org.slf4j.Logger;
+import session.Session;
 import util.FileReader;
 import util.LoggerUtil;
 import util.RequestContext;
