+++ src/main/java/session/SessionManager.java
@@ -1,6 +1,5 @@
-package repository;
+package session;
 
-import handler.Session;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
